crypto_price_automation/requirements.txt
functions-framework==3.5.0
requests
pandas
google-cloud-bigquery
google-cloud-storage
pyarrow
